--- src/main/java/org/FinalProyect_CamachoG/tempCodeRunnerFile.java
@@ -1,2 +1,0 @@
-
-        University u = new University();
