+++ src/main/java/org/example/Main.java
@@ -1,4 +1,8 @@
-package main.java.org.FinalProyect_CamachoG;
+package main.java.org.example;
+
+import main.java.org.example.imports.Initializations;
+import main.java.org.example.models.universities.University;
+import main.java.org.example.utils.Menu;
 
 public class Main {
     public static void main(String[] args) {
