+++ src/main/java/org/example/models/courses/Course.java
@@ -1,4 +1,7 @@
-package main.java.org.FinalProyect_CamachoG;
+package main.java.org.example.models.courses;
+
+import main.java.org.example.models.people.Student;
+import main.java.org.example.models.people.Teacher;
 
 import java.util.List;
