+++ src/main/java/org/example/models/people/Person.java
@@ -1,4 +1,4 @@
-package main.java.org.FinalProyect_CamachoG;
+package main.java.org.example.models.people;
 
 public abstract class Person {
