+++ src/main/java/org/example/models/universities/University.java
@@ -1,4 +1,8 @@
-package main.java.org.FinalProyect_CamachoG;
+package main.java.org.example.models.universities;
+
+import main.java.org.example.models.courses.Course;
+import main.java.org.example.models.people.Student;
+import main.java.org.example.models.people.Teacher;
 
 import java.util.List;
